backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-chess==1.999
gigachat>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
